added support for 64-bit constants
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@1529 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinxexpryy.cpp
@@ -102,7 +102,7 @@
 #if ! defined (YYSTYPE) && ! defined (YYSTYPE_IS_DECLARED)
 
 typedef union YYSTYPE {
-	int				iConst;			// constant value
+	int64_t			iConst;			// constant value
 	float			fConst;			// constant value
 	int				iAttrLocator;	// attribute locator (rowitem for int/float; offset+size for bits)
 	int				iFunc;			// function id
